+++ jupyterhub_singleuser_profiles/ui_config.py
@@ -1,10 +1,9 @@
 from pydantic import BaseModel, ValidationError, validator
-from typing import Union, Dict, Any
+from typing import Union, Dict, Any, Optional
 import json
 import logging
 
 _LOGGER = logging.getLogger(__name__)
-DEFAULT_FREQ_KEYS = ["AWS_ACCESS_KEY_ID", "AWS_SECRET_ACCESS_KEY"] # This should go to jupyterhub_config?
 
 class GpuCheckbox(BaseModel):
     value: int = 1
@@ -17,31 +16,34 @@ class GpuInput(BaseModel):
     limit: int = 1
 
 class GpuConfig(BaseModel):
-    enabled: bool = True
-    type: str
+    enabled: Optional[bool] = True
+    type: Optional[str] = None
+    gpuCheckbox: Optional[GpuCheckbox]
+    gpuDropdown: Optional[GpuDropdown]
+    gpuInput: Optional[GpuInput]
 
     @classmethod
     def create(cls, dict_:Dict[str, Any]) -> "GpuConfig":
-        v = dict_.get("config")
-        if v is None and dict_.get('type') is not None:
-            raise ValueError("No config supplied")
         if dict_.get('type') is None:
             v = None
             return cls(enabled=dict_.get('enabled'))
         if dict_['type'] == 'checkbox':
-            v = GpuCheckbox(**v)
+            v = GpuCheckbox(**dict_.get('gpuCheckbox', {}))
         elif dict_['type'] == 'dropdown':
-            v = GpuDropdown(**v)
+            v = GpuDropdown(**dict_.get('gpuDropdown', {}))
         elif dict_['type'] == 'input':
-            v = GpuInput(**v)
+            v = GpuInput(**dict_.get('gpuInput', {}))
         else:
             raise ValueError(f"Unknown type {dict_['type']}")
+        if v == {} and dict_.get('type') is not None:
+            raise ValueError("No config supplied")
         instance = cls(enabled=dict_.get('enabled'), type=dict_.get('type'), config=v)
         return instance
 
 class ImageConfig(BaseModel):
     blacklist: list = []
-    sort: str
+    whitelist: list = []
+    sort: Optional[str] = None
     #Should not be used, can lock a user out.
     #enabled: bool = True
 
@@ -53,38 +55,42 @@ class ImageConfig(BaseModel):
             raise ValueError('Sort type \"%s\" invalid' % v)
 
 class SizeConfig(BaseModel):
-    enabled: bool = True
+    enabled: Optional[bool] = True
 
 class EnvVarConfig(BaseModel):
-    freq_keys: list = DEFAULT_FREQ_KEYS
-    enabled: bool = True
+    freq_keys: Optional[list] = []
+    enabled: Optional[bool] = True
 
 class UIConfigModel(BaseModel):
-    gpuConfig: GpuConfig = {}
-    imageConfig: ImageConfig = {}
-    sizeConfig: SizeConfig = {}
-    envVarConfig: EnvVarConfig = {}
+    gpuConfig: Optional[GpuConfig] = {}
+    imageConfig: Optional[ImageConfig] = {}
+    sizeConfig: Optional[SizeConfig] = {}
+    envVarConfig: Optional[EnvVarConfig] = {}
 
 class UIConfig():
 
-    def __init__(self, ui_cfg, default_freq_keys=[]):
-        self.default_freq_keys = default_freq_keys
+    def __init__(self, ui_cfg):
         self.ui_cfg = ui_cfg
     
     def validate_ui_cm(self):
-        
-        try:
-            for key, value in self.ui_cfg.items():
-                if key == "gpuConfig":
-                    value = GpuConfig.create(value)
-                if key == "imageConfig":
-                    value = ImageConfig(**value)
-                if key == "sizeConfig":
-                    value = SizeConfig(**value)
-                if key == "envVarConfig":
-                    value = EnvVarConfig(**value)
-        except ValidationError as e:
-            _LOGGER.error("UI ConfigMap validation failed! %s" % e)
-            return json.dumps({})
+        # Unmodified or empty ui_config is a list istead of dict
+        if type(self.ui_cfg) == dict:
+            try:
+                for key, value in self.ui_cfg.items():
+                    if value is None:
+                        value = {}
+                    if key == "gpuConfig":
+                        value = GpuConfig.create(value)
+                    if key == "imageConfig":
+                        value = ImageConfig(**value)
+                    if key == "sizeConfig":
+                        value = SizeConfig(**value)
+                    if key == "envVarConfig":
+                        value = EnvVarConfig(**value)
+            except ValidationError as e:
+                _LOGGER.error("UI ConfigMap validation failed! %s" % e)
+                return {}
+        else:
+            return {}
         ui = UIConfigModel(**self.ui_cfg)
         return ui.dict()